Declare the default chuck size for uploading numeric vectors. #220
Numeric vectors will be broken into chucks of size 10^7 while being uploaded to SciDB. This size has been chosen becuase during tests it was discovered that POST-ing to SciDB via shim has a data size limit. This limit for integer vectors was around 10^9 but for double and long double this limit can be significantly lower. Maximum vector size for double or double with long precision is not entirely confirmed hence setting it to a moderately safe size.