digest_auth future-proofing
- digest_auth, lines 244-245: rearranged parentheses for sprintf to exclude the algo parameter. more info below.

algo is listed as a parameter within sprintf, and not as a parameter in digest. In R<4.0, sprintf allowed '...' parameters, but wouldn't do anything with those extra parameters. However, in R>=4.0, sprintf became more strict, only allowing the same number of parameters as there were substitutions in the given string.

Why this issue matters is, if you have R>=4.0 and want to set up a local instance, running scidbconnect would fail because the response would be the very warning message that sprintf gives you for adding more parameters than substitutions. In turn, you can't establish a proper local connection to scidb.